sample shift to overwrite existing reconnection sample
