+++ device/iot-device-samples/device-reconnection-sample/src/main/java/samples/com/microsoft/azure/sdk/iot/DeviceClientManagerSample.java
@@ -17,11 +17,8 @@ import java.util.concurrent.*;
 
 
 @Slf4j
-public class DeviceClientManagerSample {
-
-    final static List<String> failedMessageListOnClose = new ArrayList<>(); // List of messages that failed on close
-    private static DeviceClientManager deviceClientManager;
-
+public class DeviceClientManagerSample
+{
     /**
      * Sends a number of messages to an IoT. Default protocol is to use AMQP transport.
      */
@@ -61,74 +58,12 @@ public class DeviceClientManagerSample {
         }
         log.debug("Setup parameter: Protocol = [{}]", protocol);
 
-        int argNumRequest = Integer.parseInt(params.getNumRequests());
-        log.debug("Setup parameter: Requests = [{}]", argNumRequest);
-        int argSleepDuration = Integer.parseInt(params.getSleepDuration());
-        log.debug("Setup parameter: Sleep Duration = [{}]", argSleepDuration);
-        int argTimeout = Integer.parseInt(params.getTimeout());
-        log.debug("Setup parameter: Timeout = [{}]", argTimeout);
-
         DeviceClient client = new DeviceClient(argDeviceConnectionString, argProtocol);
         log.info("Successfully created an IoT Hub client.");
 
-        deviceClientManager = new DeviceClientManager(client);
-        deviceClientManager.setOperationTimeout(argTimeout);
+        DeviceClientManager deviceClientManager = new DeviceClientManager(client);
 
-        deviceClientManager.open();
-        log.info("Opened connection to IoT Hub.");
-
-        log.debug("Setting C2D message handler...");
-        deviceClientManager.setMessageCallback(new SampleMessageReceiveCallback(), new Object());
-
-        log.debug("Start sending telemetry ...");
-        startSendingTelemetry(argNumRequest, argSleepDuration);
-
-        // close the connection
-        log.info("Closing");
-        deviceClientManager.close();
-
-        if (! failedMessageListOnClose.isEmpty()) {
-            log.error("List of messages that were cancelled on close:");
-            log.error(failedMessageListOnClose.toString());
-        }
-
-        log.info("Shutting down...");
-    }
-
-    private static void startSendingTelemetry(int numRequest, int sleepTime) {
-        log.debug("Sending the following event messages:");
-        for (int i = 0; i < numRequest; ++i) {
-            Message msg = composeMessage(i);
-            SampleMessageSendCallback callback = new SampleMessageSendCallback();
-            try {
-                deviceClientManager.sendEventAsync(msg, callback, null);
-            } catch (Exception e) {
-                failedMessageListOnClose.add(msg.getMessageId());
-                log.error("Exception thrown while sending telemetry: ", e);
-            }
-            try {
-                log.debug("Sleeping for {} secs before sending next message.", sleepTime);
-                TimeUnit.SECONDS.sleep(sleepTime);
-            } catch (InterruptedException e) {
-                log.error("Exception thrown while sleeping: ", e);
-            }
-        }
-    }
-
-    private static Message composeMessage(int counter) {
-        double temperature;
-        double humidity;
-
-        temperature = 20 + Math.random() * 10;
-        humidity = 30 + Math.random() * 20;
-        String messageId = java.util.UUID.randomUUID().toString();
-
-        String msgStr = String.format(">> {\"count\": %d, \"messageId\": %s, \"temperature\": %f, \"humidity\": %f}", counter, messageId, temperature, humidity);
-        Message msg = new Message(msgStr);
-        msg.setProperty("temperatureAlert", temperature > 28 ? "true" : "false");
-        msg.setMessageId(messageId);
-        log.debug(msgStr);
-
-        return msg;
+        log.info("Starting IoT Hub client...");
+        deviceClientManager.run();
     }
 }
